requirements.txt
requests
python-dotenv
psycopg[binary]>=3.2
SQLAlchemy>=2.0
celery>=5.3
flower>=2.0
redis>=5.0